@@ -10,9 +10,14 @@ use Laminas\Diactoros\Response\EmptyResponse;
 use Laminas\Diactoros\Response\HtmlResponse;
 use Laminas\Diactoros\Response\JsonResponse;
 use Laminas\Diactoros\Response\TextResponse;
+use Laminas\HttpHandlerRunner\Emitter\HeadersSent;
 use Laminas\HttpHandlerRunner\Emitter\SapiStreamEmitter;
+use Laminas\HttpHandlerRunner\Exception\EmitterException;
 use LaminasTest\HttpHandlerRunner\TestAsset\HeaderStack;
 use LaminasTest\HttpHandlerRunner\TestAsset\MockStreamHelper;
+use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\TestCase;
+use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\StreamInterface;
 
 use function gc_collect_cycles;
@@ -26,6 +31,7 @@ use function ob_end_clean;
 use function ob_end_flush;
 use function ob_get_clean;
 use function ob_start;
+use function sprintf;
 use function str_repeat;
 use function strlen;
 use function substr;
@@ -33,22 +39,111 @@ use function substr;
 /**
  * @psalm-import-type ParsedRangeType from SapiStreamEmitter
  */
-class SapiStreamEmitterTest extends AbstractEmitterTest
+class SapiStreamEmitterTest extends TestCase
 {
-    public function setUp(): void
+    private SapiStreamEmitter $emitter;
+
+    protected function setUp(): void
     {
         HeaderStack::reset();
+        HeadersSent::reset();
         $this->emitter = new SapiStreamEmitter();
+    }
+
+    protected function tearDown(): void
+    {
+        HeaderStack::reset();
+        HeadersSent::reset();
+    }
+
+    public function testEmitsResponseHeaders(): void
+    {
+        $response = (new Response())
+            ->withStatus(200)
+            ->withAddedHeader('Content-Type', 'text/plain');
+        $response->getBody()->write('Content!');
+
+        ob_start();
+        $this->emitter->emit($response);
+        ob_end_clean();
+
+        self::assertTrue(HeaderStack::has('HTTP/1.1 200 OK'));
+        self::assertTrue(HeaderStack::has('Content-Type: text/plain'));
+    }
+
+    public function testEmitsMessageBody(): void
+    {
+        $response = (new Response())
+            ->withStatus(200)
+            ->withAddedHeader('Content-Type', 'text/plain');
+        $response->getBody()->write('Content!');
+
+        $this->expectOutputString('Content!');
+        $this->emitter->emit($response);
+    }
+
+    public function testMultipleSetCookieHeadersAreNotReplaced(): void
+    {
+        $response = (new Response())
+            ->withStatus(200)
+            ->withAddedHeader('Set-Cookie', 'foo=bar')
+            ->withAddedHeader('Set-Cookie', 'bar=baz');
+
+        $this->emitter->emit($response);
+
+        $expectedStack = [
+            ['header' => 'Set-Cookie: foo=bar', 'replace' => false, 'status_code' => 200],
+            ['header' => 'Set-Cookie: bar=baz', 'replace' => false, 'status_code' => 200],
+            ['header' => 'HTTP/1.1 200 OK', 'replace' => true, 'status_code' => 200],
+        ];
+
+        self::assertSame($expectedStack, HeaderStack::stack());
+    }
+
+    public function testDoesNotLetResponseCodeBeOverriddenByPHP(): void
+    {
+        $response = (new Response())
+            ->withStatus(202)
+            ->withAddedHeader('Location', 'http://api.my-service.com/12345678')
+            ->withAddedHeader('Content-Type', 'text/plain');
+
+        $this->emitter->emit($response);
+
+        $expectedStack = [
+            ['header' => 'Location: http://api.my-service.com/12345678', 'replace' => true, 'status_code' => 202],
+            ['header' => 'Content-Type: text/plain', 'replace' => true, 'status_code' => 202],
+            ['header' => 'HTTP/1.1 202 Accepted', 'replace' => true, 'status_code' => 202],
+        ];
+
+        self::assertSame($expectedStack, HeaderStack::stack());
+    }
+
+    public function testWillThrowEmitterExceptionWhenHeadersAreAlreadySent(): void
+    {
+        $sentInLine = __LINE__;
+        HeadersSent::markSent(__FILE__, $sentInLine);
+
+        $this->expectException(EmitterException::class);
+        $this->expectExceptionMessage(
+            sprintf(
+                'Unable to emit response; headers already sent in %s:%d',
+                __FILE__,
+                $sentInLine
+            )
+        );
+        $this->emitter->emit($this->createMock(ResponseInterface::class));
     }
 
     public function testEmitCallbackStreamResponse(): void
     {
+        $emitter  = new SapiStreamEmitter();
         $stream   = new CallbackStream(static fn(): string => 'it works');
         $response = (new Response())
             ->withStatus(200)
             ->withBody($stream);
         ob_start();
-        $this->emitter->emit($response);
+        $result = $emitter->emit($response);
+        self::assertTrue($result);
         self::assertSame('it works', ob_get_clean());
     }
 
@@ -76,7 +171,7 @@ class SapiStreamEmitterTest extends AbstractEmitterTest
     /**
      * @psalm-return array<array-key, array{0: bool, 1: bool, 2: string, 3: int}>
      */
-    public function emitStreamResponseProvider(): array
+    public static function emitStreamResponseProvider(): array
     {
         return [
             [true,   true,    '01234567890987654321',   10],
@@ -107,12 +202,12 @@ class SapiStreamEmitterTest extends AbstractEmitterTest
     }
 
     /**
-     * @dataProvider emitStreamResponseProvider
      * @param bool    $seekable        Indicates if stream is seekable
      * @param bool    $readable        Indicates if stream is readable
      * @param string  $contents        Contents stored in stream
      * @param int     $maxBufferLength Maximum buffer length used in the emitter call.
      */
+    #[DataProvider('emitStreamResponseProvider')]
     public function testEmitStreamResponse(bool $seekable, bool $readable, string $contents, int $maxBufferLength): void
     {
         $startPosition    = 0;
@@ -188,7 +283,7 @@ class SapiStreamEmitterTest extends AbstractEmitterTest
      *     4: int
      * }>
      */
-    public function emitRangeStreamResponseProvider(): array
+    public static function emitRangeStreamResponseProvider(): array
     {
         return [
             // seekable, readable,                   range,                 contents, max buffer length
@@ -244,7 +339,6 @@ class SapiStreamEmitterTest extends AbstractEmitterTest
     }
 
     /**
-     * @dataProvider emitRangeStreamResponseProvider
      * @param bool   $seekable        Indicates if stream is seekable
      * @param bool   $readable        Indicates if stream is readable
      * @param array  $range           Emitted range of data [$unit, $first, $last, $length]
@@ -252,6 +346,7 @@ class SapiStreamEmitterTest extends AbstractEmitterTest
      * @param int    $maxBufferLength Maximum buffer length used in the emitter call.
      * @psalm-param ParsedRangeType $range
      */
+    #[DataProvider('emitRangeStreamResponseProvider')]
     public function testEmitRangeStreamResponse(
         bool $seekable,
         bool $readable,
@@ -345,7 +440,7 @@ class SapiStreamEmitterTest extends AbstractEmitterTest
      *     5: int
      * }>
      */
-    public function emitMemoryUsageProvider(): array
+    public static function emitMemoryUsageProvider(): array
     {
         return [
             // seekable, readable,  size,  max,      range, max-length
@@ -377,7 +472,6 @@ class SapiStreamEmitterTest extends AbstractEmitterTest
     }
 
     /**
-     * @dataProvider emitMemoryUsageProvider
      * @param bool       $seekable         Indicates if stream is seekable
      * @param bool       $readable         Indicates if stream is readable
      * @param int        $sizeBlocks       Number the blocks of stream data.
@@ -387,6 +481,7 @@ class SapiStreamEmitterTest extends AbstractEmitterTest
      * @param int        $maxBufferLength  Maximum buffer length used in the emitter call.
      * @psalm-param array{0:int,1:int}|null $rangeBlocks
      */
+    #[DataProvider('emitMemoryUsageProvider')]
     public function testEmitMemoryUsage(
         bool $seekable,
         bool $readable,
@@ -536,7 +631,7 @@ class SapiStreamEmitterTest extends AbstractEmitterTest
     /**
      * @psalm-return array<array-key, array>
      */
-    public function emitJsonResponseProvider(): array
+    public static function emitJsonResponseProvider(): array
     {
         // @codingStandardsIgnoreStart
         return [
@@ -552,9 +647,9 @@ class SapiStreamEmitterTest extends AbstractEmitterTest
     }
 
     /**
-     * @dataProvider emitJsonResponseProvider
      * @param mixed $contents Contents stored in stream
      */
+    #[DataProvider('emitJsonResponseProvider')]
     public function testEmitJsonResponse(mixed $contents): void
     {
         $response = (new JsonResponse($contents))
@@ -582,7 +677,7 @@ class SapiStreamEmitterTest extends AbstractEmitterTest
     /**
      * @psalm-return array<array-key, array{0: string, 1: string, 2: string}>
      */
-    public function contentRangeProvider(): array
+    public static function contentRangeProvider(): array
     {
         return [
             ['bytes 0-2/*', 'Hello world', 'Hel'],
@@ -591,9 +686,7 @@ class SapiStreamEmitterTest extends AbstractEmitterTest
         ];
     }
 
-    /**
-     * @dataProvider contentRangeProvider
-     */
+    #[DataProvider('contentRangeProvider')]
     public function testContentRange(string $header, string $body, string $expected): void
     {
         $response = (new Response())
